fix(app-5nu): Т42a фикс-раунд 1 — перенос видимости при смене точки обзора, арифметика кулдауна, доки
I-1 (критично, подтверждено координатором): принятое переключение точки обзора не
сбрасывало per-connection память видимости (VisibilitySet Previous/Current, латч Р133) —
гистерезис/линжер SnapshotAssembler ещё LingerTicks тиков отдавали живые координаты со
СТАРОЙ точки, сводя на нет смысл кулдауна Р70. Новый SnapshotAssembler.ResetViewpointMemory
(наборы видимости + латч; подписки на снаряды осознанно не трогаются — память личности,
не точки обзора) зовётся из MatchServer.OnSpectateRequest в ветке принятия.

I-2 (дефект брифа координатора, подтверждено независимым измерением через dotnet):
(double)float*int в конверсии секунд в тики — ровно ловушка Р141: на 0.1/0.2/0.4 при
TickRate 20/30/60 даёт на тик больше, чем точное значение (28-46 из 120 проверенных
комбинаций расходятся с ceil от decimal.Parse исходного литерала). Заменено на
(decimal)float*int — 0 расхождений на той же матрице.

I-3: класс-док/отчёт утверждали позитивный витнес у каждого негативного теста — у двух
(TargetIsSelf, Order_RequesterAliveWins) витнеса не было. Добавлены оба (второй заодно
закрывает половину M-2).

I-4: предикат диапазона targetIndex жил в двух домах (SpectatePolicy.Evaluate и
MatchServer.OnSpectateRequest) — вынесен в SpectatePolicy.IsTargetInRange, оба места
зовут одну копию.

M-1 (ещё один дефект брифа): обоснование порядка CooldownActive-последним было
контрфактично (кулдаун тратится только на принятие) — переписано на настоящую причину
(диагностика в логе). M-2: добавлены два кейса на порядок проверок. M-4: маркер-комментарий
NetConfig приведён из будущего времени в прошедшее. M-5: класс-док MatchServer описывает
оба permanent-корня (OnPostTick + SpectateRequestNet), не один. M-6: лог отказа
троттлится по смене причины на слот (не по тику), с сбросом памяти на принятое
переключение.

Новый тест SnapshotAssemblerTests (линжер/гистерезис + позитивный витнес), два новых
теста SpectateTests (порядок TargetDead/CooldownActive, IsTargetInRange), 8 мутаций с
откатом по md5 (два control-мутанта в непокрытой юнитами MatchServer/ServerBootstrap
проводке — честно не пойманы, включая мутацию из самого брифа на снятие вызова сброса,
что и есть архитектурная граница этого класса, не дефект). Полный прогон 751/751, оба
golden поимённо Passed, DeterminismTests.cs без изменений.

+++ client/Assets/Scripts/Server/ServerBootstrap.cs
@@ -321,16 +321,43 @@ namespace Ring.Server
             // `SpectatePolicy` (Stage 2 Task 42a) takes the same kind of
             // finished tick count, built the same way and for the same
             // reason (`SpectatePolicy`'s own doc). The rounding is UP
-            // (`Math.Ceiling`), never a bare `(int)` truncation: at the
-            // shipped default `0.35 * 30 = 10.5`, and truncating would enforce
-            // a 10-tick (0.333s) cooldown — SHORTER than the 0.35s the asset
-            // names, which weakens the exact limiter Р70 exists to enforce.
-            // Ceiling gives 11 ticks (0.367s) instead — never shorter than
-            // configured, only ever equal or longer.
+            // (`Math.Ceiling`), never a bare `(int)` truncation: truncating
+            // could enforce a cooldown SHORTER than the asset names, which
+            // weakens the exact limiter Р70 exists to enforce.
+            //
+            // `decimal`, NOT `double` (Task 42a fix-round 1, I-2 — this was
+            // the round's own defect, since corrected; the earlier comment
+            // here claimed "0.35 * 30 = 10.5", which is a statement about the
+            // MATHEMATICAL real numbers, not about what this line actually
+            // computes). `(double)SpectatorSwitchCooldownSeconds * TickRate`
+            // is exactly the trap spec Р141 names: widening a `float` to
+            // `double` preserves the float's OWN rounding error instead of
+            // fixing it, so a value the Inspector shows as a clean decimal
+            // can land fractionally ABOVE its intended product — measured
+            // directly (not assumed): `(double)0.1f * 30` is
+            // `3.0000000447034836`, whose `Ceiling` is `4`, one tick more
+            // than the `3` the number `0.1` actually means; the same defect
+            // reproduces at `0.2` (`Ceiling` `7`, not `6`) and `0.4`
+            // (`Ceiling` `13`, not `12`), silently lengthening the cooldown
+            // by exactly one tick on those and roughly a third of the legal
+            // `[Range(0.05f, 2f)]` × `{20, 30, 60}` combinations checked
+            // while fixing this. The shipped default `0.35` happens to be
+            // unaffected purely by chance (`10.499999821186066` still ceils
+            // to `11`, the same answer the exact `10.5` would) — which is
+            // exactly why R-TEST alone, run only against the shipped
+            // default, could never have caught this. `(decimal)float`
+            // widens through the float's OWN ~7-significant-digit precision
+            // instead of `double`'s ~15-17, which is enough to recover the
+            // clean decimal a human actually typed into the Inspector for
+            // every value checked above — verified against the
+            // mathematically exact product computed straight from the
+            // decimal LITERAL (`decimal.Parse` on the same text, never
+            // touching `float` at all), not merely against `double`'s own
+            // answer.
             try
             {
                 int spectateCooldownTicks = (int)Math.Ceiling(
-                    (double)_net.SpectatorSwitchCooldownSeconds * _net.TickRate);
+                    (decimal)_net.SpectatorSwitchCooldownSeconds * _net.TickRate);
                 _matchServer = new MatchServer(_nm, _net,
                     new MatchEndPolicy(_net.MatchMaxDurationSeconds * _net.TickRate),
                     new SpectatePolicy(spectateCooldownTicks));
